requirements.txt
opencv-python==4.11.0.86
pip==25.0
scikit-learn==1.6.1
tensorflow==2.15.0
tensorflow-hub==0.13.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.67.1
